take a little more care with memory allocations

+++ src/lib/codec/image_format/IFileIO.h
@@ -26,7 +26,7 @@ struct GrkIOBuf : public grk_io_buf
 {
   public:
 	GrkIOBuf() : GrkIOBuf(nullptr, 0, 0, 0, false) {}
-	GrkIOBuf(uint8_t* data, uint64_t offset, uint64_t dataLen, uint64_t allocLen, bool pooled)
+	GrkIOBuf(uint8_t* data, size_t offset, size_t dataLen, size_t allocLen, bool pooled)
 	{
 		data_ = data;
 		offset_ = offset;
@@ -42,7 +42,7 @@ struct GrkIOBuf : public grk_io_buf
 		allocLen_ = rhs.allocLen_;
 		pooled_ = rhs.pooled_;
 	}
-	bool alloc(uint64_t len)
+	bool alloc(size_t len)
 	{
 		dealloc();
 		data_ = (uint8_t*)grk_bin::grk_aligned_malloc(len);
